+++ CHANGELOG.md
@@ -10,7 +10,7 @@ We go to the next version after each release on the
 ### Added
 
 - Airplane Mode Toggle Button Visibility in Quick Settings.
-- More Crypto Donations (USDC, LTC, Shiba Inu and LBRY).
+- More Crypto Donations (XRP, Solana, Cardano, BNB, Stellar, USDC, LTC, Shiba Inu and LBRY).
 - Support Notifier.
 
 ## [31.0] - 2025-01-06